+++ .ebextensions/options.config
@@ -1,5 +1,5 @@
 option_settings:
   aws:elasticbeanstalk:application:environment:
-    FLASK_APP: app
+    FLASK_APP: application.py
     FLASK_ENV: production
     AWS_DEFAULT_REGION: us-east-1
